requirements.txt
beautifulsoup4
selenium
webdriver-manager
python-telegram-bot==20.3
